Move tests to proper namespace; enforce signatures
